Refactored and organized file names for pantry, list, and notification-related components
