+++ app/src/main/java/com/example/restock/pantry/PantryAdapter.java
@@ -37,7 +37,7 @@ public class PantryAdapter extends RecyclerView.Adapter<PantryAdapter.ViewHolder
     @NonNull
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.item_pantry, parent, false);
+                .inflate(R.layout.pantry_item, parent, false);
         return new ViewHolder(view);
     }
