--- app/src/main/res/layout/fragment_list_list.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!-- fragment_list_list.xml -->
-<androidx.recyclerview.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/list"
-    android:name="com.example.restock.ListFragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_marginLeft="16dp"
-    android:layout_marginRight="16dp"
-    app:layoutManager="LinearLayoutManager"
-    tools:context=".ListFragment"
-    tools:listitem="@layout/item_list" />
