+++ app/src/main/res/layout/notification_item.xml
@@ -7,7 +7,7 @@
     android:padding="16dp"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:background="@drawable/white_box_bg"
+    android:background="@drawable/white_rounded_box_bg"
     android:elevation="2dp"
     android:layout_marginBottom="8dp">
