+++ app/src/main/res/layout/pantry_fragment.xml
@@ -45,7 +45,7 @@
         android:layout_height="50dp"
         android:background="@drawable/custom_search_bar"
         android:orientation="horizontal"
-        android:padding="8dp"
+        android:padding="12dp"
         app:layout_constraintTop_toBottomOf="@id/pantry_title"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
@@ -64,12 +64,12 @@
         <EditText
             android:id="@+id/search_input"
             android:layout_width="0dp"
-            android:layout_height="50dp"
+            android:layout_height="48dp"
             android:layout_weight="1"
             android:hint="@string/search_hint"
             android:paddingHorizontal="10dp"
             android:textSize="16sp"
-            android:background="@null"
+            android:textColor="@color/black"
             android:inputType="text"
             android:autofillHints="search"
             android:gravity="center_vertical"/>
@@ -100,7 +100,7 @@
         app:layout_constraintBottom_toTopOf="@id/footer_navigation"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        tools:listitem="@layout/item_pantry"/>
+        tools:listitem="@layout/pantry_item"/>
 
     <!-- Divider Above Footer -->
     <View
