+++ app/src/main/res/layout/pantry_item.xml
